+++ week03/kyulim/java-baseball-6/src/main/java/baseball/controller/BaseballGame.java
@@ -3,7 +3,6 @@ package baseball.controller;
 import baseball.model.BaseballReferee;
 import baseball.view.InputView;
 import baseball.view.OutputView;
-import camp.nextstep.edu.missionutils.Randoms;
 
 import java.util.*;
 
@@ -20,7 +19,6 @@ public class BaseballGame {
         this.inputView = new InputView();
     }
 
-    // 게임 세팅
     public void settingGame() {
         outputView.printStartGame();
         playGame();
@@ -29,12 +27,10 @@ public class BaseballGame {
             settingGame();
     }
 
-    // 게임 시작
     public void playGame() {
-        List<Integer> computerNumbers = generateComputerNumbers();
+        List<Integer> computerNumbers = baseballReferee.generateComputerNumbers();
         while (true) {
-            String input = getUserInput();
-            List<Integer> userNumbers = convertInputType(input);
+            List<Integer> userNumbers = getUserNumbers();
 
             String result = baseballReferee.getResult(computerNumbers, userNumbers);
             outputView.printResult(result);
@@ -46,24 +42,18 @@ public class BaseballGame {
         }
     }
 
-    // 컴퓨터 숫자 랜덤으로 정하기
-    public List<Integer> generateComputerNumbers() {
-        Set<Integer> numberSet = new HashSet<>();
-        while (numberSet.size() < 3) {
-            numberSet.add(Randoms.pickNumberInRange(1, 9));
-        }
-        return new ArrayList<>(numberSet);
+    private List<Integer> getUserNumbers() {
+        String input = getUserInput();
+        return convertInputType(input);
     }
 
-    // 사용자 숫자 입력받기
     public String getUserInput() {
         outputView.printInputNumber();
         String input = inputView.inputNumber();
-        validateInput(input);
+        baseballReferee.validateInput(input);
         return input;
     }
 
-    // 재시작 여부 묻기
     public boolean askRestart() {
         outputView.printRestartGame();
         String input = inputView.inputNumber();
@@ -74,7 +64,6 @@ public class BaseballGame {
         return input.equals("1");
     }
 
-    // 사용자 숫자 타입 변경 (String -> List<Integer>)
     public List<Integer> convertInputType(String input) {
         List<Integer> numbers = new ArrayList<>();
         for (String num : input.split(""))
@@ -82,28 +71,4 @@ public class BaseballGame {
 
         return numbers;
     }
-
-    // 사용자 숫자 입력 형식 확인
-    public void validateInput(String input) {
-        if (input.length() != 3) {
-            throw new IllegalArgumentException("잘못된 입력입니다.");
-        }
-
-        for (char ch : input.toCharArray()) {
-            if (ch < '1' || ch > '9') {
-                throw new IllegalArgumentException("잘못된 입력입니다.");
-            }
-        }
-
-        if (hasDuplicateCharacters(input)) {
-            throw new IllegalArgumentException("잘못된 입력입니다.");
-        }
-    }
-
-    // 사용자 숫자 입력 중복 확인(HashSet: 증복 허용X)
-    public boolean hasDuplicateCharacters(String input) {
-        return input.length() != new HashSet<>(Arrays.asList(input.split(""))).size();
-    }
-
-
 }
