+++ week03/kyulim/java-baseball-6/src/main/java/baseball/model/BaseballReferee.java
@@ -1,18 +1,30 @@
 package baseball.model;
 
-import java.util.List;
+import camp.nextstep.edu.missionutils.Randoms;
+
+import java.util.*;
 
 public class BaseballReferee {
-    // 볼, 스트라이크 출력
+
+    public List<Integer> generateComputerNumbers() {
+        Set<Integer> numberSet = new HashSet<>();
+        while (numberSet.size() < 3) {
+            numberSet.add(Randoms.pickNumberInRange(1, 9));
+        }
+        return new ArrayList<>(numberSet);
+    }
+
     public String getResult(List<Integer> computerNumbers, List<Integer> userNumbers) {
         int strikes = countStrikes(computerNumbers, userNumbers);
         int balls = countBalls(computerNumbers, userNumbers);
 
-        String result = "";
-        if (strikes == 0 && balls == 0) {
-            result = "낫싱";
-        }
+        return getNothingOrBallOrStrike(strikes, balls);
+    }
 
+    private String getNothingOrBallOrStrike(int strikes, int balls) {
+        String result = "";
+        if (strikes == 0 && balls == 0)
+            result = "낫싱";
         if (balls > 0)
             result += balls + "볼 ";
         if (strikes > 0)
@@ -21,7 +33,6 @@ public class BaseballReferee {
         return result.trim();
     }
 
-    // 스트라이크 개수 정하기
     private int countStrikes(List<Integer> computerNumber, List<Integer> userNumber) {
         int strikes = 0;
         for (int i = 0; i < 3; i++) {
@@ -31,7 +42,6 @@ public class BaseballReferee {
         return strikes;
     }
 
-    // 볼 개수 정하기
     private int countBalls(List<Integer> computerNumber, List<Integer> userNumber) {
         int balls = 0;
         for (int i = 0; i < 3; i++) {
@@ -42,5 +52,25 @@ public class BaseballReferee {
                 balls--;
         }
         return balls;
+    }
+
+    public void validateInput(String input) {
+        if (input.length() != 3) {
+            throw new IllegalArgumentException("잘못된 입력입니다.");
+        }
+
+        for (char ch : input.toCharArray()) {
+            if (ch < '1' || ch > '9') {
+                throw new IllegalArgumentException("잘못된 입력입니다.");
+            }
+        }
+
+        if (hasDuplicateCharacters(input)) {
+            throw new IllegalArgumentException("잘못된 입력입니다.");
+        }
+    }
+
+    public boolean hasDuplicateCharacters(String input) {
+        return input.length() != new HashSet<>(Arrays.asList(input.split(""))).size();
     }
 }
